migration to ila_swagger packages
