--- ila/lib/swagger/api/account_api.dart
@@ -1,163 +1,0 @@
-part of swagger.api;
-
-
-
-class AccountApi {
-  final ApiClient apiClient;
-
-  AccountApi([ApiClient apiClient]) : apiClient = apiClient ?? defaultApiClient;
-
-  /// 
-  ///
-  /// 
-  Future<MultipartFile> accountLogout() async {
-    Object postBody = null;
-
-    // verify required params are set
-
-    // create path and map variables
-    String path = "/api/Account/logout".replaceAll("{format}","json");
-
-    // query params
-    List<QueryParam> queryParams = [];
-    Map<String, String> headerParams = {};
-    Map<String, String> formParams = {};
-    
-    List<String> contentTypes = [];
-
-    String contentType = contentTypes.length > 0 ? contentTypes[0] : "application/json";
-    List<String> authNames = ["JWT"];
-
-    if(contentType.startsWith("multipart/form-data")) {
-      bool hasFields = false;
-      MultipartRequest mp = new MultipartRequest(null, null);
-      
-      if(hasFields)
-        postBody = mp;
-    }
-    else {
-          }
-
-    var response = await apiClient.invokeAPI(path,
-                                             'POST',
-                                             queryParams,
-                                             postBody,
-                                             headerParams,
-                                             formParams,
-                                             contentType,
-                                             authNames);
-
-    if(response.statusCode >= 400) {
-      throw new ApiException(response.statusCode, response.body);
-    } else if(response.body != null) {
-      return 
-          apiClient.deserialize(response.body, 'MultipartFile') as MultipartFile ;
-    } else {
-      return null;
-    }
-  }
-  /// 
-  ///
-  /// 
-  Future<JsonWebToken> accountSignIn(SignIn request) async {
-    Object postBody = request;
-
-    // verify required params are set
-    if(request == null) {
-     throw new ApiException(400, "Missing required param: request");
-    }
-
-    // create path and map variables
-    String path = "/api/Account/login".replaceAll("{format}","json");
-
-    // query params
-    List<QueryParam> queryParams = [];
-    Map<String, String> headerParams = {};
-    Map<String, String> formParams = {};
-    
-    List<String> contentTypes = ["application/json-patch+json","application/json","text/json","application/_*+json"];
-
-    String contentType = contentTypes.length > 0 ? contentTypes[0] : "application/json";
-    List<String> authNames = [];
-
-    if(contentType.startsWith("multipart/form-data")) {
-      bool hasFields = false;
-      MultipartRequest mp = new MultipartRequest(null, null);
-      
-      if(hasFields)
-        postBody = mp;
-    }
-    else {
-          }
-
-    var response = await apiClient.invokeAPI(path,
-                                             'POST',
-                                             queryParams,
-                                             postBody,
-                                             headerParams,
-                                             formParams,
-                                             contentType,
-                                             authNames);
-
-    if(response.statusCode >= 400) {
-      throw new ApiException(response.statusCode, response.body);
-    } else if(response.body != null) {
-      return 
-          apiClient.deserialize(response.body, 'JsonWebToken') as JsonWebToken ;
-    } else {
-      return null;
-    }
-  }
-  /// 
-  ///
-  /// 
-  Future<MultipartFile> accountSignUp(SignUp request) async {
-    Object postBody = request;
-
-    // verify required params are set
-    if(request == null) {
-     throw new ApiException(400, "Missing required param: request");
-    }
-
-    // create path and map variables
-    String path = "/api/Account/signup".replaceAll("{format}","json");
-
-    // query params
-    List<QueryParam> queryParams = [];
-    Map<String, String> headerParams = {};
-    Map<String, String> formParams = {};
-    
-    List<String> contentTypes = ["application/json-patch+json","application/json","text/json","application/_*+json"];
-
-    String contentType = contentTypes.length > 0 ? contentTypes[0] : "application/json";
-    List<String> authNames = [];
-
-    if(contentType.startsWith("multipart/form-data")) {
-      bool hasFields = false;
-      MultipartRequest mp = new MultipartRequest(null, null);
-      
-      if(hasFields)
-        postBody = mp;
-    }
-    else {
-          }
-
-    var response = await apiClient.invokeAPI(path,
-                                             'POST',
-                                             queryParams,
-                                             postBody,
-                                             headerParams,
-                                             formParams,
-                                             contentType,
-                                             authNames);
-
-    if(response.statusCode >= 400) {
-      throw new ApiException(response.statusCode, response.body);
-    } else if(response.body != null) {
-      return 
-          apiClient.deserialize(response.body, 'MultipartFile') as MultipartFile ;
-    } else {
-      return null;
-    }
-  }
-}
