--- ila/lib/swagger/auth/authentication.dart
@@ -1,7 +1,0 @@
-part of swagger.api;
-
-abstract class Authentication {
-
-	/// Apply authentication settings to header and query params.
-	void applyToParams(List<QueryParam> queryParams, Map<String, String> headerParams);
-}
