+++ ila/pubspec.yaml
@@ -23,10 +23,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-  http: any
   scoped_model: any
-  jaguar_jwt: any
   flutter_secure_storage: any
+  ila_swagger:
+    path: ../ila_swagger/
 
 dev_dependencies:
   flutter_test:
