+++ ila_swagger/lib/api_client.dart
@@ -8,6 +8,14 @@ class QueryParam {
 }
 
 class ApiClient {
+  String token;
+  String username;
+  String password;
+  JWT jwt;
+
+  AuthStatus _status = AuthStatus.None;
+
+  AuthStatus getStatus() => _status;
 
   String basePath;
   var client = new Client();
@@ -19,12 +27,11 @@ class ApiClient {
   final _RegMap = new RegExp(r'^Map<String,(.*)>$');
 
   ApiClient({this.basePath: "https://ila.dev.kirschbaum.bayern"}) {
-    // Setup authentications (key: authentication name, value: authentication).
-    _authentications['JWT'] = new ApiKeyAuth("header", "Authorization");
+    _authentications['JWT'] = new ApiKeyAuth("header", "Authorization", "Bearer");
   }
 
   void addDefaultHeader(String key, String value) {
-     _defaultHeaderMap[key] = value;
+    _defaultHeaderMap[key] = value;
   }
 
   dynamic _deserialize(dynamic value, String targetType) {
@@ -84,9 +91,11 @@ class ApiClient {
           }
       }
     } catch (e, stack) {
-      throw new ApiException.withInner(500, 'Exception during deserialization.', e, stack);
+      throw new ApiException.withInner(
+          500, 'Exception during deserialization.', e, stack);
     }
-    throw new ApiException(500, 'Could not find a suitable class for deserialization');
+    throw new ApiException(
+        500, 'Could not find a suitable class for deserialization');
   }
 
   dynamic deserialize(String jsonVal, String targetType) {
@@ -112,27 +121,45 @@ class ApiClient {
   // We don't use a Map<String, String> for queryParams.
   // If collectionFormat is 'multi' a key might appear multiple times.
   Future<Response> invokeAPI(String path,
-                             String method,
-                             Iterable<QueryParam> queryParams,
-                             Object body,
-                             Map<String, String> headerParams,
-                             Map<String, String> formParams,
-                             String contentType,
-                             List<String> authNames) async {
+      String method,
+      Iterable<QueryParam> queryParams,
+      Object body,
+      Map<String, String> headerParams,
+      Map<String, String> formParams,
+      String contentType,
+      List<String> authNames) async {
+
+    if (authNames.length != 0) {
+      // token renewal
+      if (jwt != null &&
+          jwt.expiresAt
+              .subtract(Duration(seconds: 30))
+              .difference(DateTime.now().toUtc())
+              .isNegative) {
+        print('update JWT');
+        try {
+          await login(username, password);
+        } catch(error){
+          // TODO: redirect to login
+        }
+        print('JWT updated');
+      }
+    }
 
     _updateParamsForAuth(authNames, queryParams, headerParams);
 
-    var ps = queryParams.where((p) => p.value != null).map((p) => '${p.name}=${p.value}');
+    var ps = queryParams.where((p) => p.value != null).map((p) => '${p.name}=${p
+        .value}');
     String queryString = ps.isNotEmpty ?
-                         '?' + ps.join('&') :
-                         '';
+    '?' + ps.join('&') :
+    '';
 
     String url = basePath + path + queryString;
 
     headerParams.addAll(_defaultHeaderMap);
     headerParams['Content-Type'] = contentType;
 
-    if(body is MultipartRequest) {
+    if (body is MultipartRequest) {
       var request = new MultipartRequest(method, Uri.parse(url));
       request.fields.addAll(body.fields);
       request.files.addAll(body.files);
@@ -141,8 +168,10 @@ class ApiClient {
       var response = await client.send(request);
       return Response.fromStream(response);
     } else {
-      var msgBody = contentType == "application/x-www-form-urlencoded" ? formParams : serialize(body);
-      switch(method) {
+      var msgBody = contentType == "application/x-www-form-urlencoded"
+          ? formParams
+          : serialize(body);
+      switch (method) {
         case "POST":
           return client.post(url, headers: headerParams, body: msgBody);
         case "PUT":
@@ -159,19 +188,99 @@ class ApiClient {
 
   /// Update query and header parameters based on authentication settings.
   /// @param authNames The authentications to apply
-  void _updateParamsForAuth(List<String> authNames, List<QueryParam> queryParams, Map<String, String> headerParams) {
+  void _updateParamsForAuth(List<String> authNames,
+      List<QueryParam> queryParams, Map<String, String> headerParams) {
     authNames.forEach((authName) {
       Authentication auth = _authentications[authName];
-      if (auth == null) throw new ArgumentError("Authentication undefined: " + authName);
+      if (auth == null) throw new ArgumentError(
+          "Authentication undefined: " + authName);
       auth.applyToParams(queryParams, headerParams);
     });
   }
 
   void setAccessToken(String accessToken) {
     _authentications.forEach((key, auth) {
-      if (auth is OAuth) {
-        auth.setAccessToken(accessToken);
+      if (auth is ApiKeyAuth) {
+        auth.apiKey = accessToken;
       }
     });
   }
+
+  Future login(String username, String password) async {
+    _status = AuthStatus.Login;
+    try {
+      var response = await AccountApi().accountSignIn(
+          SignIn.fromJson({'username': username, 'password': password}));
+
+      final decoded =
+      B64urlEncRfc7515.decodeUtf8(response.accessToken.split('.')[1]);
+      jwt = JWT.fromJsonString(jsonDecode(decoded));
+      token = response.accessToken;
+
+      this.username = username;
+      this.password = password;
+      defaultApiClient.setAccessToken(token);
+
+      _status = AuthStatus.LoggedIn;
+    } catch (error) {
+      _status = AuthStatus.LoginFailed;
+
+      this.username = null;
+      this.password = null;
+      jwt = null;
+      this.setAccessToken(null);
+
+      if (error is ApiException) {
+        throw UserException(message: error.message);
+      } else {
+        throw UserException(message: 'Undefined Error occured');
+      }
+    }
+  }
+
+  logout() async {
+    try {
+      await AccountApi().accountLogout();
+      username = null;
+      password = null;
+      jwt = null;
+      _status = AuthStatus.None;
+      token = null;
+      setAccessToken(null);
+    } catch (error) {
+      if (error is ApiException) {
+        throw UserException(message: error.message);
+      } else {
+        throw UserException(message: 'Undefined Error occured');
+      }
+    }
+  }
+}
+
+class JWT {
+  int id;
+  int iat;
+  int exp;
+  String sub;
+  String aud;
+  String iss;
+  DateTime expiresAt;
+
+  JWT.fromJsonString(Map<String, dynamic> json)
+      : id = json['id'],
+        iat = json['iat'],
+        exp = json['exp'],
+        sub = json['sub'],
+        aud = json['aud'],
+        iss = json['iss'] {
+    expiresAt = DateTime.fromMillisecondsSinceEpoch(exp * 1000);
+  }
+}
+
+enum AuthStatus {
+  None,
+  Login,
+  LoggedIn,
+  LoginFailed,
+  // Expired,
 }
