+++ ila_swagger/lib/userException.dart
@@ -1,3 +1,5 @@
+part of ila_swagger.api;
+
 class UserException {
   final String message;
