+++ ila_swagger/pubspec.yaml
@@ -3,6 +3,7 @@ version: 1.0.0
 description: Swagger API client
 dependencies:
   http: '>=0.11.1 <0.12.0'
+  jaguar_jwt: any
 
 environment:
   sdk: ">=2.1.0 <3.0.0"
